% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_colours.R
\docType{data}
\name{sci_colours}
\alias{sci_colours}
\title{SCI Colours}
\format{
An object of class \code{character} of length 35.
}
\usage{
sci_colours
}
\description{
Internal function containing the info of the different SCI colours
}
\keyword{datasets}
